requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.10
idna==3.3
iso8601==1.0.2
loguru==0.5.3
m3u8==1.0.0
powerline-status==2.7
pycryptodome==3.12.0
requests==2.27.1
urllib3==1.26.8
